backend/requirements.txt
pytesseract
transformers
absl-py==2.0.0
aes-everywhere==1.2.10
anyio==4.1.0
APScheduler==3.10.4
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
crccheck==1.3.1
crispy-bootstrap5==0.7
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.0
daphne==4.0.0
deepface==0.0.84
Django==4.2.2
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-mqtt==0.4.0
django-picklefield==3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djcompiler==0.0.5
drf-yasg==1.20.0
exceptiongroup==1.2.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.3
fpdf==1.7.2
fsspec==2024.5.0
gast==0.4.0
gdown==4.7.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.23.1
hyperlink==21.0.0
idna==3.4
imbalanced-learn==0.12.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
imutils==0.5.4
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
JIT==0.2.7
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.10.0
kiwisolver==1.4.5
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.7
mtcnn==0.1.1
mysqlclient==2.2.4
narwhals==1.32.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
oauthlib==3.2.2
openapi==1.1.0
openapi-codec==1.3.2
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
pdfkit==1.0.0
Pillow==10.0.0
plotly==6.0.1
protobuf==3.19.6
pyarmor.cli.core==4.3.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodomex==3.18.0
pydyf==0.11.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyphen==0.16.0
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWave==0.5.0
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.4.2
redis==5.0.1
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soundfile==0.13.1
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tbb==2021.12.0
tensorflow==2.10.0
termcolor==2.3.0
threadpoolctl==3.3.0
tinycss2==1.4.0
tokenizers==0.13.3
txaio==23.1.1
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.24.0.post1
vedirect==1.1.0
waitress==2.1.2
Wave==0.0.2
wcwidth==0.2.6
weasyprint==61.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wfdb==4.1.2
whitenoise==6.6.0
zipp==3.16.2
zope.interface==6.0
zopfli==0.2.3.post1
torch
numpy==1.24.3
pdf2image
pytesseract